spark/core, spark/lib: [Fix] Fix new GCC 13 errors (#15)

+++ spark/core/src/platforms/SfmlWindow.cpp
@@ -94,7 +94,7 @@ namespace spark::core
                 }
             default:
                 {
-                    SPARK_CORE_WARN("Unhandled event type: {}", event.type);
+                    SPARK_CORE_WARN("Unhandled event type: {}", static_cast<int>(event.type));
                     break;
                 }
             }
